Craigslist/requirements.txt
browser-use>=0.1.20
python-dotenv>=1.0.1
google-generativeai>=0.7.0
flask>=3.0.0
openai>=1.40.0
